id validation done

+++ controllers/user.controllers.js
@@ -23,10 +23,11 @@ const addUser = (req, res) => {
 	fs.readFile('user.json', (err, data) => {
 		if (err) return res.status(500).send({ success: false, message: err });
 		const userData = JSON.parse(data);
-		console.log(userData.length);
-		const { gender, name, contact, address, photoURL } = req.body;
+		const { gender, name, contact, address, photoURL, id } = req.body;
+		
+
 		const newUserInfo = {
-			id: userData.length + 1,
+			id: id,
 			gender,
 			name,
 			contact,
@@ -35,11 +36,15 @@ const addUser = (req, res) => {
 		};
 
 		// validation
+		if (!id) return res.status(400).send({ success: false, message: 'id is required' });
 		if (!gender) return res.status(400).send({ success: false, message: 'gender is required' });
 		if (!name) return res.status(400).send({ success: false, message: 'name is required' });
 		if (!contact) return res.status(400).send({ success: false, message: 'contact is required' });
 		if (!address) return res.status(400).send({ success: false, message: 'address is required' });
 		if (!photoURL) return res.status(400).send({ success: false, message: 'photoURL is required' });
+
+		const isIdExists = userData.find((user) => user.id === id);
+		if (isIdExists) return res.status(400).send({ success: false, message: 'id already exists' });
 
 		fs.writeFile('user.json', JSON.stringify([...userData, newUserInfo]), (err) => {
 			if (err) return res.status(500).send({ success: false, message: err });
@@ -48,8 +53,16 @@ const addUser = (req, res) => {
 	});
 };
 
+const updateSingleUser = (req, res) => {
+	const { id } = req.params.id;
+	if (!id) return res.status(400).send({ success: false, message: 'gender is required' });
+	console.log(id);
+	res.send(id);
+};
+
 module.exports = {
 	getAllUsers,
 	getRandomUser,
 	addUser,
+	updateSingleUser,
 };
